fix upload image

+++ src/main/java/com/biblio/service/impl/PublisherServiceImpl.java
@@ -13,11 +13,14 @@ import com.biblio.enumeration.EBookTemplateStatus;
 import com.biblio.enumeration.EOrderStatus;
 import com.biblio.mapper.PublisherMapper;
 import com.biblio.service.IPublisherService;
+import com.biblio.utils.ManageFileUtil;
 
 import javax.inject.Inject;
+import javax.transaction.Transactional;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 public class PublisherServiceImpl implements IPublisherService {
     @Inject
@@ -92,14 +95,51 @@ public class PublisherServiceImpl implements IPublisherService {
     }
 
     @Override
-    public void updatePublisher(PublisherUpdateRequest publisherUpdateRequest) {
-        Publisher publisher = publisherDAO.getEntityById(Long.valueOf(publisherUpdateRequest.getId()));
-        publisherDAO.updatePublisher(PublisherMapper.toPublisherUpdate(publisherUpdateRequest, publisher));
+
+    public Boolean updatePublisher(PublisherUpdateRequest publisherUpdateRequest) {
+        try {
+            Publisher publisher = publisherDAO.getEntityById(Long.valueOf(publisherUpdateRequest.getId()));
+            if (publisher == null) {
+                throw new Exception("Publisher not found with ID: " + publisherUpdateRequest.getId());
+            }
+
+            if (!Objects.equals(publisher.getAvatar(), publisherUpdateRequest.getAvatar())) {
+                Boolean isAvatarDeleted = ManageFileUtil.deleteFileAvatar(publisher.getAvatar(), "publisher");
+                if (!isAvatarDeleted) {
+                    throw new Exception("Failed to delete avatar for publisher: " + publisherUpdateRequest.getId());
+                }
+            }
+
+            publisherDAO.updatePublisher(PublisherMapper.toPublisherUpdate(publisherUpdateRequest, publisher));
+
+            return true;
+        } catch (Exception e) {
+            System.err.println("Transaction failed: " + e.getMessage());
+            return false;
+        }
     }
 
     @Override
-    public void deletePublisher(PublisherDeleteRequest publisherDeleteRequest) {
-        publisherDAO.deletePublisher(Long.valueOf(publisherDeleteRequest.getId()));
+    @Transactional
+    public Boolean deletePublisher(PublisherDeleteRequest publisherDeleteRequest) {
+        try {
+            Publisher publisher = publisherDAO.getEntityById(Long.valueOf(publisherDeleteRequest.getId()));
+            if (publisher == null) {
+                throw new Exception("Publisher not found with ID: " + publisherDeleteRequest.getId());
+            }
+
+            Boolean isAvatarDeleted = ManageFileUtil.deleteFileAvatar(publisher.getAvatar(), "publisher");
+            if (!isAvatarDeleted) {
+                throw new Exception("Failed to delete avatar for publisher: " + publisherDeleteRequest.getId());
+            }
+
+            publisherDAO.deletePublisher(Long.valueOf(publisherDeleteRequest.getId()));
+
+            return true;
+        } catch (Exception e) {
+            System.err.println("Transaction failed: " + e.getMessage());
+            return false;
+        }
     }
 
     @Override
